refactor: settings service layer
