+++ src/features/settings/data/settings.mutations.ts
@@ -1,12 +1,11 @@
 import { useMutation } from "@tanstack/react-query";
-import { api } from "@/libs/api.lib";
-import { Settings, SettingsResponseSchema } from "./settings.schema";
+import { Settings } from "./settings.schemas";
+import { createSettings } from "./settings.services";
 
 export const useCreateSettingsMutation = () => {
   return useMutation({
     async mutationFn(data: Settings) {
-      const res = await api.post(data, "/settings");
-      return SettingsResponseSchema.parse(res);
+      return await createSettings(data);
     },
   });
 };
