+++ src/features/settings/data/settings.queries.ts
@@ -1,16 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { api } from "@/libs/api.lib";
-import { SettingsResponseSchema } from "./settings.schema";
-
-export const getSettings = async () => {
-  const res = await api.get("/settings");
-  return SettingsResponseSchema.parse(res);
-};
-
-export const preloadSettings = async () => {
-  const settings = await getSettings();
-  window.settings = settings.data;
-};
+import { getSettings } from "./settings.services";
 
 export const useSettingsQuery = () =>
   useQuery({
